+++ Final/instructions21.py
@@ -5,6 +5,7 @@ from gtts import gTTS
 import queue, threading, subprocess
 
 global instructionProcess
+instructionProcess = ""
 outputInstruct21 =  queue.Queue() 
 def outPutAudioWarInstruct21():
     global outputInstruct21, instructionProcess
@@ -22,6 +23,7 @@ def instructions21(audioChoice):
     root21I.title('Instructions 21')
     root21I['background']='#8B0000'
     root21I.geometry("1000x1000")
+    root21I.state("zoomed")
     instructionsGeneral = "Rules of 21:\n \
     Cards values of 2-9 are worth 2points-9pointts, respectively.\n\
     King, Jack, and Queen are each worth 10points.\n \
@@ -62,14 +64,15 @@ def instructions21(audioChoice):
     If the player or dealer have the same score, that defines a push, which is a tie.\n \
     Press the hit button to receive a card. Press the stand button to stand.\n\
     If playing with audio, at any time say 'yes' to hit, say 'no' to stand, and say 'score' for the score. To quit say 'quit' or close out window."
-    placeGeneral = Label(root21I, text = instructionsGeneral, font=("Comic Sans MS", 12),  bg='#8B0000')
+    placeGeneral = Label(root21I, text = instructionsGeneral, font=("Tahoma", 18), bg='#8B0000', foreground="white")
     placeGeneral.place(relx= 0, rely=0)
-    placeApplication = Label(root21I, text = instructionsApplication, font=("Comic Sans MS", 12),  bg='#8B0000')
-    placeApplication.place(relx= 0, rely=.6)
+    placeApplication = Label(root21I, text = instructionsApplication, font=("Tahoma", 18), bg='#8B0000', foreground="white")
+    placeApplication.place(relx= 0, rely=.8)
     if audioChoice:
         outputInstruct21.put(instructionString)
         audioSpeakerThread = threading.Thread(target=outPutAudioWarInstruct21)
         audioSpeakerThread.start()
     root21I.mainloop()
-    instructionProcess.terminate()
+    if instructionProcess != "":
+        instructionProcess.terminate()
