+++ Final/instructionsWar.py
@@ -5,6 +5,7 @@ from gtts import gTTS
 import queue, threading, subprocess
 
 global outPutAudioWarInstructWar, instructionProcess
+instructionProcess = ""
 outputInstructWar =  queue.Queue() 
 
 def outPutAudioWarInstructWar(rootWarI):
@@ -23,6 +24,7 @@ def instructionsWar(audioChoice):
     rootWarI.title('Instructions War')
     rootWarI['background']='#8B0000'
     rootWarI.geometry("1000x1000")
+    rootWarI.state("zoomed")
     instructionString = "Rules of War: \
     The deck is split evenly. The player and computer flips a card and the higher card gets both cards.\
     If there is a tie, the computer and player flip four cards and the fourth card is compared.\
@@ -43,9 +45,10 @@ def instructionsWar(audioChoice):
     instructionsApplication = "With Press the flip button to flip your card (the computer will flip their card too).\n \
     If playing with audio say flip to do the same. When a war commences, say war to continue or hit the war button. \n\
     At any point, if playing with audio say score to see how many cards the player and computer have each."
-    placeGeneral = Label(rootWarI, text = instructionsGeneral, font=("Comic Sans MS", 12),  bg='#8B0000')
+    placeGeneral = Label(rootWarI, text = instructionsGeneral,font=("Tahoma", 18), bg='#8B0000', foreground="white")
     placeGeneral.place(relx= 0, rely=0)
-    placeApplication = Label(rootWarI, text = instructionsApplication, font=("Comic Sans MS", 12),  bg='#8B0000')
+    placeApplication = Label(rootWarI, text = instructionsApplication,font=("Tahoma", 18), bg='#8B0000', foreground="white")
     placeApplication.place(relx= 0, rely=.6)
     rootWarI.mainloop()
-    instructionProcess.terminate()
+    if instructionProcess != "":
+        instructionProcess.terminate()
